Правки к сервисам доступа к данным
Скорректировал сервисы для прохождения тестов. Есть общая логика, необходимо будет вывести все в отдельный абстрактный класс(сократит количества кода и тестов )

+++ src/main/java/com/test/carfines/service/impl/CarBrandServiceImpl.java
@@ -1,26 +1,52 @@
 package com.test.carfines.service.impl;
 
 import com.test.carfines.model.CarBrand;
-
 import com.test.carfines.repository.CarBrandRepository;
 import com.test.carfines.service.CarBrandService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 
+@Service
+@Slf4j
 public class CarBrandServiceImpl implements CarBrandService {
+    private final CarBrandRepository repository;
 
     @Autowired
-    private CarBrandRepository repository;
+    public CarBrandServiceImpl(CarBrandRepository repository) {
+        this.repository = repository;
+    }
+
 
     @Override
-    public CarBrand addCarBrand(CarBrand carBrand) {
-        return repository.saveAndFlush( carBrand );
+    public boolean addCarBrand(CarBrand carBrand) {
+        var carBrandFromDB = repository.findByName( carBrand.getCarBrandName() );
+        if (carBrandFromDB != null) {
+            return false;
+        }
+
+        repository.saveAndFlush( carBrand );
+        log.info( "Added " + carBrand.toString() + " " + LocalDate.now() );
+        return true;
     }
 
     @Override
-    public void delete(long id) {
-         repository.deleteById( id );
+    public boolean delete(long id) {
+        Optional<CarBrand> carBrandFromDB = repository.findById( id );
+
+        if(carBrandFromDB == null){
+            return false;
+        }
+
+        repository.deleteById( id );
+
+        log.info( "Delete " + carBrandFromDB.toString() + " " + LocalDate.now() );
+
+        return true;
     }
 
     @Override
@@ -29,8 +55,14 @@ public class CarBrandServiceImpl implements CarBrandService {
     }
 
     @Override
-    public CarBrand editCarBrand(CarBrand carBrand) {
-        return repository.saveAndFlush( carBrand );
+    public boolean editCarBrand(CarBrand carBrand) {
+        Optional<CarBrand> carBrandFromDB = repository.findById( carBrand.getId() );
+        if (carBrandFromDB == null) {
+            return false;
+        }
+        repository.saveAndFlush( carBrand );
+        log.info( "Edit  old version = " + carBrandFromDB.toString() + ", new version =" + carBrand.toString() + " " + LocalDate.now() );
+        return true;
     }
 
     @Override
